🐛 fix(recally): address code review issues
- Split 995-line recally.go into recally.go, recally_articles.go,
  recally_feeds.go, recally_digest.go
- Remove dead UpsertArticle query whose is_new detection via
  created_at = updated_at was unreliable; regenerate sqlc
- CreateFeedEntry returns nil,nil on ErrNoRows (duplicate) so poll
  command can distinguish dedup from real DB errors
- Guard buildFrontmatter against metadata keys overwriting reserved
  fields (id, url, title, status, etc.)
- Add NOT NULL to articles.id in schema source
- Check json.MarshalIndent errors throughout recally CLI commands
- Skip anna binary copy in EnsureAnnaCLIInPath when size matches
