+++ pkg/db/sqlc/articles.sql.go
@@ -553,114 +553,3 @@ func (q *Queries) UpdateArticle(ctx context.Context, arg UpdateArticleParams) (A
 	)
 	return i, err
 }
-
-const upsertArticle = `-- name: UpsertArticle :one
-INSERT INTO articles (
-    id, user_id, agent_id, url, canonical_url, source_type,
-    title, author, summary, tags, status, starred, file_path, metadata,
-    published_at, saved_at, read_at
-)
-VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-ON CONFLICT(user_id, canonical_url) DO UPDATE SET
-    url           = excluded.url,
-    agent_id      = excluded.agent_id,
-    source_type   = excluded.source_type,
-    title         = excluded.title,
-    author        = excluded.author,
-    summary       = excluded.summary,
-    tags          = excluded.tags,
-    file_path     = excluded.file_path,
-    metadata      = excluded.metadata,
-    published_at  = excluded.published_at,
-    updated_at    = datetime('now')
-RETURNING id, user_id, agent_id, url, canonical_url, source_type, title, author, summary, tags, status, starred, file_path, metadata, published_at, saved_at, read_at, created_at, updated_at, (CASE WHEN created_at = updated_at THEN 1 ELSE 0 END) AS is_new
-`
-
-type UpsertArticleParams struct {
-	ID           string         `json:"id"`
-	UserID       int64          `json:"user_id"`
-	AgentID      sql.NullString `json:"agent_id"`
-	Url          string         `json:"url"`
-	CanonicalUrl string         `json:"canonical_url"`
-	SourceType   string         `json:"source_type"`
-	Title        string         `json:"title"`
-	Author       string         `json:"author"`
-	Summary      string         `json:"summary"`
-	Tags         string         `json:"tags"`
-	Status       string         `json:"status"`
-	Starred      int64          `json:"starred"`
-	FilePath     string         `json:"file_path"`
-	Metadata     string         `json:"metadata"`
-	PublishedAt  sql.NullString `json:"published_at"`
-	SavedAt      string         `json:"saved_at"`
-	ReadAt       sql.NullString `json:"read_at"`
-}
-
-type UpsertArticleRow struct {
-	Column1      interface{}    `json:"column_1"`
-	ID           string         `json:"id"`
-	UserID       int64          `json:"user_id"`
-	AgentID      sql.NullString `json:"agent_id"`
-	Url          string         `json:"url"`
-	CanonicalUrl string         `json:"canonical_url"`
-	SourceType   string         `json:"source_type"`
-	Title        string         `json:"title"`
-	Author       string         `json:"author"`
-	Summary      string         `json:"summary"`
-	Tags         string         `json:"tags"`
-	Status       string         `json:"status"`
-	Starred      int64          `json:"starred"`
-	FilePath     string         `json:"file_path"`
-	Metadata     string         `json:"metadata"`
-	PublishedAt  sql.NullString `json:"published_at"`
-	SavedAt      string         `json:"saved_at"`
-	ReadAt       sql.NullString `json:"read_at"`
-	CreatedAt    string         `json:"created_at"`
-	UpdatedAt    string         `json:"updated_at"`
-}
-
-func (q *Queries) UpsertArticle(ctx context.Context, arg UpsertArticleParams) (UpsertArticleRow, error) {
-	row := q.db.QueryRowContext(ctx, upsertArticle,
-		arg.ID,
-		arg.UserID,
-		arg.AgentID,
-		arg.Url,
-		arg.CanonicalUrl,
-		arg.SourceType,
-		arg.Title,
-		arg.Author,
-		arg.Summary,
-		arg.Tags,
-		arg.Status,
-		arg.Starred,
-		arg.FilePath,
-		arg.Metadata,
-		arg.PublishedAt,
-		arg.SavedAt,
-		arg.ReadAt,
-	)
-	var i UpsertArticleRow
-	err := row.Scan(
-		&i.Column1,
-		&i.ID,
-		&i.UserID,
-		&i.AgentID,
-		&i.Url,
-		&i.CanonicalUrl,
-		&i.SourceType,
-		&i.Title,
-		&i.Author,
-		&i.Summary,
-		&i.Tags,
-		&i.Status,
-		&i.Starred,
-		&i.FilePath,
-		&i.Metadata,
-		&i.PublishedAt,
-		&i.SavedAt,
-		&i.ReadAt,
-		&i.CreatedAt,
-		&i.UpdatedAt,
-	)
-	return i, err
-}
